hal: realtek: bee: update rtl87x2g drivers
Update Realtek Bee series rtl87x2g drivers.
Sync from: Ie1d792d0fe53b40c65d2f1ade1b13f4264077e05

+++ bee/drivers/keyscan/src/device/rtl87x2g/rtl_keyscan.c
@@ -282,7 +282,7 @@ void KeyScan_FilterDataConfig(KEYSCAN_TypeDef *KeyScan, uint16_t data, Functiona
   *         This parameter can be: ENABLE or DISABLE.
   * \return none.
   */
-void KeyScan_debounceConfig(KEYSCAN_TypeDef *KeyScan, uint8_t time,
+void KeyScan_debounceConfig(KEYSCAN_TypeDef *KeyScan, uint16_t time,
                             FunctionalState NewState)
 {
     /* Check the parameters */
